Moved the 'query_tree::print' function implementation over to a source file.
